+++ src/components/MenuBar/index.tsx
@@ -1,8 +1,7 @@
 interface MenuBarProps {
   setPage: (page: number) => void;
 }
-
-const icons = [
+const LapnoteIcon = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     className="h-6 w-6"
@@ -18,7 +17,10 @@ const icons = [
         
 2.599-.333-2.599-1  .895-1 2.599-1 2.599.333 2.599 1z"
     />
-  </svg>,
+  </svg>
+);
+
+const BookshelfIcon = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     className="h-6 w-6"
@@ -32,7 +34,27 @@ const icons = [
       strokeWidth={2}
       d="M12 6v6m0 0v6m0-6h6m-6 0H6"
     />
-  </svg>,
+  </svg>
+);
+
+const LogbookIcon = (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    className="h-6 w-6"
+    fill="none"
+    viewBox="0 0 24 24"
+    stroke="currentColor"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
+    />
+  </svg>
+);
+
+const StatsIcon = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     className="h-6 w-6"
@@ -46,7 +68,10 @@ const icons = [
       strokeWidth={2}
       d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"
     />
-  </svg>,
+  </svg>
+);
+
+const SettingsIcon = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     className="h-6 w-6"
@@ -66,7 +91,14 @@ const icons = [
       strokeWidth={2}
       d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
     />
-  </svg>,
+  </svg>
+);
+const icons = [
+  LapnoteIcon,
+  BookshelfIcon,
+  LogbookIcon,
+  StatsIcon,
+  SettingsIcon,
 ];
 
 const MenuBar: React.FC<MenuBarProps> = ({ setPage }) => {
