+++ src/pages/lapnote/TimerControls.tsx
