+++ src/pages/lapnote/TimerDisplay.tsx
